\name{Barlow2014}
\alias{Barlow2014}

\title{A Meta-Analysis on the Effect of Parent Training Programs vs. Control for Improving Parental Psychosocial Health Within 4 Weeks After Intervention}
\description{This data set contains 26 studies with the standardized mean differences between the improvement in parental psychosocial health for subjects who were enrolled in parent training programs vs. those who were not. This data set is also available in the \code{R} package \code{altmeta}, along with many other useful data sets.
 }
\usage{data(Barlow2014)}
\format{
A dataframe with 26 studies with the following five variables within each study.
\describe{
    \item{\code{y}:}{standardized mean differences in improvement of parental psychosocial health.}
    \item{\code{s}:}{sample standard errors of standardized mean differences.}
    \item{\code{n1}:}{sample sizes in treatment group 1 (parent training programs).}
    \item{\code{n2}:}{sample sizes in treatment group 2 (control).}
    \item{\code{n}:}{total sample size.}
  }
}

\source{ 
Barlow J, Smailagic N, Huband N, Roloff V, Bennett C (2014). "Group-based parent training programmes for improving parental psychosocial health." \emph{Cochrane Database of Systematic Reviews},\bold{5}, Art. No.: CD002020. <doi: 10.1002/14651858.CD002020.pub4>
}